git-svn-id: https://svn.apache.org/repos/asf/commons/proper/net/tags/NET_2_0@696100 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/net/time/TimeTCPClient.java
@@ -81,7 +81,7 @@ public final class TimeTCPClient extends SocketClient
     {
         DataInputStream input;
         input = new DataInputStream(_input_);
-        return (long)(input.readInt() & 0xffffffffL);
+        return (input.readInt() & 0xffffffffL);
     }
 
     /***
